Add files via upload
Here I run a very simple comparative analysis of percolation between two, different kind of lattice networks:
1. Square
2. Square but including diagonals

I show that the more pathways we include in the network, the more robust it will become against random failure of nodes (in this particular case) for different thresholds of failure. In other words, the lattice with diagonals is more robust than the square lattice with no diagonals.